UPDATE: simplify files
