+++ ai/ai.py
@@ -1,7 +1,5 @@
 import pickle, os, json
-# from pong import pong_game 
-from game import pong_game 
-from utils import DISPLAY_GAME, DYSPLAY_LOG, NB_GENERATION, NB_SPECIES, SAVE_AI, SAVE_FILE, SAVE_FOLDER
+from game import pong_train, DISPLAY_GAME, DYSPLAY_LOG, NB_GENERATION, NB_SPECIES, SAVE_AI, SAVE_FILE, SAVE_FOLDER
 import numpy as np
 
 np.random.seed()
@@ -167,7 +165,7 @@ def train_Ai(save_file, base):
 
         for i in range(NB_SPECIES):
             Ai_Sample[i].ai_score = 0
-            if pong_game(Ai_Sample[i], DISPLAY_GAME) == "STOP":
+            if pong_train(Ai_Sample[i], DISPLAY_GAME) == "STOP":
                 return
             if (DYSPLAY_LOG == "yes"):
                 print(f"The AI opponent {i} send back the ball {Ai_Sample[i].ai_score} times")
