+++ ai/main.py
@@ -1,8 +1,6 @@
 import os, sys
 from ai import train_Ai, load_Ai
-from utils import SAVE_FILE, SAVE_FOLDER
-# from pong import play_Ai
-from game import play_Ai
+from game import play_Ai, SAVE_FILE, SAVE_FOLDER
 
 # Run the game
 def main():
